+++ mcumgr-toolkit-cli/src/errors.rs
@@ -49,7 +49,7 @@ pub enum CliError {
     #[error("Failed to parse MCUboot image")]
     #[diagnostic(code(mcumgrctl::image_parse))]
     ImageParseFailed(#[from] ImageParseError),
-    #[error("Firmware update failed")]
-    #[diagnostic(code(mcumgrctl::firmware_update))]
+    #[error(transparent)]
+    #[diagnostic(transparent)]
     FirmwareUpdateFailed(#[from] FirmwareUpdateError),
 }