+++ mcumgr-toolkit-cli/src/main.rs
@@ -87,9 +87,15 @@ fn cli_main(multiprogress: &MultiProgress) -> Result<(), CliError> {
     };
 
     if let Ok(client) = client.get() {
+        client.set_retries(args.retries);
+
         if let Err(e) = client.use_auto_frame_size() {
+            let mut lowest_err: &dyn std::error::Error = &e;
+            while let Some(lower_err) = lowest_err.source() {
+                lowest_err = lower_err;
+            }
             log::warn!("Failed to read SMP frame size from device, using slow default");
-            log::warn!("Reason: {e}");
+            log::warn!("Reason: {lowest_err}");
             log::warn!("Hint: Make sure that `CONFIG_MCUMGR_GRP_OS_MCUMGR_PARAMS` is enabled.");
         }
     }