+++ mcumgr-toolkit/src/connection.rs
@@ -1,18 +1,26 @@
 use std::{io::Cursor, sync::Mutex, time::Duration};
 
 use crate::{
+    DEFAULT_RETRIES,
     commands::{ErrResponse, ErrResponseV2, McuMgrCommand},
     smp_errors::{DeviceError, MCUmgrErr},
     transport::{ReceiveError, SendError, Transport},
 };
 
 use miette::{Diagnostic, IntoDiagnostic};
+use polonius_the_crab::prelude::*;
 use thiserror::Error;
 
-struct Inner {
+struct Transceiver {
     transport: Box<dyn Transport + Send>,
     next_seqnum: u8,
-    transport_buffer: Box<[u8; u16::MAX as usize]>,
+    receive_buffer: Box<[u8; u16::MAX as usize]>,
+}
+
+struct Inner {
+    transceiver: Transceiver,
+    send_buffer: Box<[u8; u16::MAX as usize]>,
+    retries: u8,
 }
 
 /// An SMP protocol layer connection to a device.
@@ -59,14 +67,77 @@ impl ExecuteError {
     }
 }
 
+impl Transceiver {
+    fn transceive_command(
+        &mut self,
+        write_operation: bool,
+        group_id: u16,
+        command_id: u8,
+        data: &[u8],
+    ) -> Result<&'_ [u8], ExecuteError> {
+        let sequence_num = self.next_seqnum;
+        self.next_seqnum = self.next_seqnum.wrapping_add(1);
+
+        self.transport
+            .send_frame(write_operation, sequence_num, group_id, command_id, data)?;
+
+        self.transport
+            .receive_frame(
+                &mut self.receive_buffer,
+                write_operation,
+                sequence_num,
+                group_id,
+                command_id,
+            )
+            .map_err(Into::into)
+    }
+
+    fn transceive_command_with_retries(
+        &mut self,
+        write_operation: bool,
+        group_id: u16,
+        command_id: u8,
+        data: &[u8],
+        num_retries: u8,
+    ) -> Result<&'_ [u8], ExecuteError> {
+        let mut this = self;
+
+        let mut counter = 0;
+
+        polonius_loop!(|this| -> Result<&'polonius [u8], ExecuteError> {
+            let result = this.transceive_command(write_operation, group_id, command_id, data);
+
+            if counter >= num_retries {
+                polonius_return!(result)
+            }
+            counter += 1;
+
+            match result {
+                Ok(_) => polonius_return!(result),
+                Err(e) => {
+                    let mut lowest_err: &dyn std::error::Error = &e;
+                    while let Some(lower_err) = lowest_err.source() {
+                        lowest_err = lower_err;
+                    }
+                    log::warn!("Retry transmission, error occurred: {lowest_err}");
+                }
+            }
+        })
+    }
+}
+
 impl Connection {
-    /// Creates a new SMP
+    /// Creates a new SMP connection
     pub fn new<T: Transport + Send + 'static>(transport: T) -> Self {
         Self {
             inner: Mutex::new(Inner {
-                transport: Box::new(transport),
-                next_seqnum: rand::random(),
-                transport_buffer: Box::new([0; u16::MAX as usize]),
+                transceiver: Transceiver {
+                    transport: Box::new(transport),
+                    next_seqnum: rand::random(),
+                    receive_buffer: Box::new([0; u16::MAX as usize]),
+                },
+                send_buffer: Box::new([0; u16::MAX as usize]),
+                retries: DEFAULT_RETRIES,
             }),
         }
     }
@@ -79,7 +150,20 @@ impl Connection {
         &self,
         timeout: Duration,
     ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-        self.inner.lock().unwrap().transport.set_timeout(timeout)
+        self.inner
+            .lock()
+            .unwrap()
+            .transceiver
+            .transport
+            .set_timeout(timeout)
+    }
+
+    /// Changes the retry amount.
+    ///
+    /// When the device encounters a transport error, it will retry
+    /// this many times until giving up.
+    pub fn set_retries(&self, retries: u8) {
+        self.inner.lock().unwrap().retries = retries;
     }
 
     /// Executes a given CBOR based SMP command.
@@ -87,40 +171,47 @@ impl Connection {
         &self,
         request: &R,
     ) -> Result<R::Response, ExecuteError> {
+        self.execute_command_impl(request, true)
+    }
+
+    /// Executes a given CBOR based SMP command.
+    ///
+    /// Does not use retries.
+    pub fn execute_command_without_retries<R: McuMgrCommand>(
+        &self,
+        request: &R,
+    ) -> Result<R::Response, ExecuteError> {
+        self.execute_command_impl(request, false)
+    }
+
+    fn execute_command_impl<R: McuMgrCommand>(
+        &self,
+        request: &R,
+        use_retries: bool,
+    ) -> Result<R::Response, ExecuteError> {
         let mut lock_guard = self.inner.lock().unwrap();
         let locked_self: &mut Inner = &mut lock_guard;
 
-        let mut cursor = Cursor::new(locked_self.transport_buffer.as_mut_slice());
+        let mut cursor = Cursor::new(locked_self.send_buffer.as_mut_slice());
         ciborium::into_writer(request.data(), &mut cursor)
             .into_diagnostic()
             .map_err(Into::into)
             .map_err(ExecuteError::EncodeFailed)?;
         let data_size = cursor.position() as usize;
-        let data = &locked_self.transport_buffer[..data_size];
+        let data = &locked_self.send_buffer[..data_size];
 
         log::debug!("TX data: {}", hex::encode(data));
-
-        let sequence_num = locked_self.next_seqnum;
-        locked_self.next_seqnum = locked_self.next_seqnum.wrapping_add(1);
 
         let write_operation = request.is_write_operation();
         let group_id = request.group_id();
         let command_id = request.command_id();
 
-        locked_self.transport.send_frame(
+        let response = locked_self.transceiver.transceive_command_with_retries(
             write_operation,
-            sequence_num,
             group_id,
             command_id,
             data,
-        )?;
-
-        let response = locked_self.transport.receive_frame(
-            &mut locked_self.transport_buffer,
-            write_operation,
-            sequence_num,
-            group_id,
-            command_id,
+            if use_retries { locked_self.retries } else { 0 },
         )?;
 
         log::debug!("RX data: {}", hex::encode(response));
@@ -166,31 +257,20 @@ impl Connection {
         group_id: u16,
         command_id: u8,
         data: &[u8],
+        use_retries: bool,
     ) -> Result<Box<[u8]>, ExecuteError> {
         let mut lock_guard = self.inner.lock().unwrap();
         let locked_self: &mut Inner = &mut lock_guard;
 
-        let sequence_num = locked_self.next_seqnum;
-        locked_self.next_seqnum = locked_self.next_seqnum.wrapping_add(1);
-
-        locked_self.transport.send_frame(
-            write_operation,
-            sequence_num,
-            group_id,
-            command_id,
-            data,
-        )?;
-
         locked_self
-            .transport
-            .receive_frame(
-                &mut locked_self.transport_buffer,
+            .transceiver
+            .transceive_command_with_retries(
                 write_operation,
-                sequence_num,
                 group_id,
                 command_id,
+                data,
+                if use_retries { locked_self.retries } else { 0 },
             )
-            .map_err(Into::into)
             .map(|val| val.into())
     }
 }