+++ mcumgr-toolkit/src/lib.rs
@@ -65,3 +65,9 @@ impl MCUmgrGroup {
         }
     }
 }
+
+/// The default timeout used by this crate
+pub const DEFAULT_TIMEOUT_MS: u64 = 500;
+
+/// The default retry count used by this crate
+pub const DEFAULT_RETRIES: u8 = 5;